+++ NEWS.md
@@ -6,6 +6,8 @@
 * Add Xie 2019 agomelatine PK model
 * Drop `qs` since it will be archived
 * Update tumor growth inhibition models
+* `addResErr()` now works with multiple-endpoint models
+* Additional testing
 
 # Version 0.3.1